requirements.txt
gradio==4.19.2
pandas>=2.0.0
requests>=2.31.0
huggingface-hub>=0.19.0
datasets>=2.15.0
openpyxl>=3.1.2
python-dotenv==1.0.0
soundfile==0.12.1
numpy==1.26.3